updated fieldToSchemaMap with custom resolvers
